Changes:
* added new "Direct Debit" payment details type
* added BankWireRef property to bank-wired payouts
* added UserId property to KycDocument entity
* added two new lists of KycDocuments (per user ID and per SDK client)
* implemented lists sorting
* removed unused API calls

+++ src/com/mangopay/core/Money.java
@@ -11,8 +11,8 @@ public class Money extends Dto {
     public String Currency;
     
     /**
-     * The currency amount of money.
+     * The currency amount of money, in cents
      */
-    public Double Amount;
+    public int Amount;
     
 }
